Gemini-private/backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI Model APIs
anthropic==0.34.0
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
vertexai==1.38.1
openai==1.6.1  # For embeddings fallback

# HTTP client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.3

# Logging and monitoring
prometheus-client==0.19.0

# LDAP for Active Directory (optional)
python-ldap==3.4.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For test client
